refactor : 재구조화
