+++ src/main/java/gift/controller/AdminController.java
@@ -6,8 +6,8 @@ import gift.dto.PagingResponse;
 import gift.model.category.CategoryResponse;
 import gift.model.gift.GiftRequest;
 import gift.model.gift.GiftResponse;
-import gift.service.CategoryService;
-import gift.service.GiftService;
+import gift.service.category.CategoryService;
+import gift.service.gift.GiftService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
