+++ src/main/java/gift/controller/UserController.java
@@ -2,7 +2,7 @@ package gift.controller;
 
 import gift.model.user.User;
 import gift.model.user.UserRequest;
-import gift.service.UserService;
+import gift.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
