+++ src/main/java/gift/controller/WishListController.java
@@ -6,8 +6,8 @@ import gift.model.gift.GiftResponse;
 import gift.model.user.User;
 import gift.model.wish.Wish;
 import gift.model.wish.WishResponse;
-import gift.service.GiftService;
-import gift.service.WishService;
+import gift.service.gift.GiftService;
+import gift.service.wish.WishService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
