+++ src/main/java/gift/model/option/OptionRequest.java
@@ -1,0 +1,29 @@
+package gift.model.option;
+
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+
+public class OptionRequest {
+
+    @Size(max = 50, message = "옵션의 이름은 최대 50자까지 입력할 수 있습니다.")
+    @Pattern(regexp = "[\\s\\(\\)\\[\\]\\+\\-&/_a-zA-Z0-9\uAC00-\uD7AF]*", message = "특수문자 오류")
+    private String name;
+
+    private int quantity;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getQuantity() {
+        return quantity;
+    }
+
+    public void setQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+}
