+++ src/main/java/gift/repository/category/CategoryRepository.java
@@ -1,4 +1,4 @@
-package gift.repository;
+package gift.repository.category;
 
 import gift.model.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
