+++ src/test/java/gift/repository/GiftRepositoryTest.java
@@ -1,13 +1,10 @@
 package gift.repository;
 
 import gift.model.gift.Gift;
-import gift.model.gift.GiftResponse;
+import gift.repository.gift.GiftRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
