+++ src/test/java/gift/repository/WishRepositoryTest.java
@@ -3,6 +3,9 @@ package gift.repository;
 import gift.model.gift.Gift;
 import gift.model.user.User;
 import gift.model.wish.Wish;
+import gift.repository.gift.GiftRepository;
+import gift.repository.user.UserRepository;
+import gift.repository.wish.WishRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
